Change location_id to Integer type

+++ app/db.py
@@ -21,7 +21,7 @@ class User(db.Model):
     bio = db.Column(db.String, nullable=False)
     email = db.Column(db.String)
     phone = db.Column(db.String)
-    location_id = db.Column(db.Float, db.ForeignKey('location.id'))
+    location_id = db.Column(db.Integer, db.ForeignKey('location.id'))
     location = db.relationship('Location')
 
     def __init__(self, **kwargs):
